Properly handle interleaved request for loading table.
It may happens that a request for loading table is sent while another
one has not yet been processed but was already scheduled. In such
case, it may happens that the tables are added twice without being
unloaded in between.

This patch fix this by unloading table before to load it in the result
handler, preventing any interleaved call.