Be tolerant to interleaved fetch request.
It may happens that fetch requests interleaved. In such case the
requested fetched rows offsets may not be exactly contiguous with the
currently loaded row window offset.

Preventing interleaved request may result in items not being shown at
all since the fetch request is not sent.

This patch no longer prevents interleaved fetch request and adjust so
insertion code to tolerate fetch range overlapping the beginning or
the end of the currently loaded rows window.

Range overlapping the entire window and non contiguous ranges are
still not supported, but should not happens.